Move function to service layer
